requirements.txt
python-dotenv
filelock
musicbrainzngs
mutagen
requests
